+++ app/src/main/java/com/woowa/banchan/ui/home/HomeActivity.kt
@@ -1,19 +1,14 @@
 package com.woowa.banchan.ui.home
 
+import com.woowa.banchan.ui.home.adapter.HomeVPAdapter
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.google.android.material.tabs.TabLayoutMediator
 import com.woowa.banchan.R
 import com.woowa.banchan.databinding.ActivityHomeBinding
 import com.woowa.banchan.ui.cart.CartActivity
-import com.woowa.banchan.ui.home.best.BestFragment
-import com.woowa.banchan.ui.home.main.MainFragment
-import com.woowa.banchan.ui.home.side.SideFragment
-import com.woowa.banchan.ui.home.soup.SoupFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -31,19 +26,7 @@ class HomeActivity : AppCompatActivity() {
     }
 
     private fun initAdapter() {
-        binding.vpHome.adapter =
-            object : FragmentStateAdapter(supportFragmentManager, lifecycle) {
-                override fun getItemCount(): Int = 4
-
-                override fun createFragment(position: Int): Fragment =
-                    when (position) {
-                        0 -> BestFragment()
-                        1 -> MainFragment()
-                        2 -> SoupFragment()
-                        3 -> SideFragment()
-                        else -> Fragment()
-                    }
-            }
+        binding.vpHome.adapter = HomeVPAdapter(supportFragmentManager, lifecycle)
     }
 
     private fun initTabLayoutMediator() {
