+++ app/src/main/java/com/woowa/banchan/ui/home/adapter/viewholder/HomeRecyclerViewViewHolder.kt
@@ -5,7 +5,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.woowa.banchan.data.remote.dto.FoodItem
 import com.woowa.banchan.databinding.ItemRecyclerviewBinding
-import com.woowa.banchan.ui.home.adapter.HomeItemAdapter
+import com.woowa.banchan.ui.home.adapter.HomeRVAdapter
 import com.woowa.banchan.utils.LINEAR_HORIZONTAL
 import com.woowa.banchan.utils.LINEAR_VERTICAL
 
@@ -13,7 +13,7 @@ class HomeRecyclerViewViewHolder(private val binding: ItemRecyclerviewBinding) :
     RecyclerView.ViewHolder(binding.root) {
 
     fun bind(food: List<FoodItem>, managerType: Int) {
-        val adapter = HomeItemAdapter(managerType)
+        val adapter = HomeRVAdapter(managerType)
         binding.layoutBest.apply {
             layoutManager = when (managerType) {
                 LINEAR_HORIZONTAL -> LinearLayoutManager(this.context, LinearLayoutManager.HORIZONTAL, false)
